refactor response contruction impl

+++ src/http/response.h
@@ -2,6 +2,8 @@
 #define HTTP_RESPONSE_H
 
 #include "static.h"
+#include "../utils/compression.h"
+#include "../utils/file.h"
 
-int construct_response(http_request_t *http_request, char *response);
+size_t construct_response(http_request_t *http_request, char *response);
 #endif // HTTP_RESPONSE_H
